답변 재생성 부분 모델, 프롬프트 분리
